### 2018-07-24 #836 @whedon
[PRE REVIEW]: Lepton
**Submitting author:** @slithiaote (<a href="http://orcid.org/0000-0002-4977-4969">Sébastien Li-Thiao-Té</a>)
**Repository:** <a href="https://github.com/slithiaote/lepton" target ="_blank">https://github.com/slithiaote/lepton</a>
**Version:** v1.0
