### 2018-07-24 #836 @jedbrown
Thanks for your submission. I have some questions and concerns:

* I don't see any source code in the repository, just an executable. Is that intentional?

* The CeCILL-B license is not OSI-approved (for the same reason 4-clause BSD is not).

https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

* Can you explain how you believe this software fits the _research_ criteria in the submission requirements?
