### 2018-07-28 #836 @slithiaote
Thank you for your comments.

- It is intentional not to have source code in a separate file from documentation :
-- In the literate programming paradigm, source code is presented in a human readable format, regardless of the constraints of the programming language. Consequently, it is usually embedded in a file in documentation format. Here, the source code is embedded in the plain text lepton.nw file, while humans should read the generated lepton.pdf file.
-- The actual machine-readable source code is intended to be extracted by Lepton from the lepton.nw file, following the instructions in Section 3 (Compilation). As Lepton generates both the pdf and the executable from these instructions, humans should not have to deal with the intermediate Ocaml .ml files.

- As far as I understand, the CeCILL licenses are listed on the OSI website as open source. Do you mean that the B version is not OSI-approved ? [https://opensource.org/licenses/CECILL-2.1](https://opensource.org/licenses/CECILL-2.1)

- I believe that Lepton fits the research criteria :
-- This software is useful for research. In our mathematics lab at Paris 13 University, it has been used to produce data analysis reports on the amounts of oxydative stress in rat eyes in response to the injection of different chemicals. Another example is that we use it to develop image analysis software for detecting ulcers in colonoscopy videos. Using Lepton guarantees that the data analysis results are produced by the source code described in the same document.
-- This software is research in itself. In the last decade, several solutions have been proposed to improve the reproducibility of research results and in particular data analysis. The current trend is "notebooks" such as Jupyter or Matlab notebooks, i.e. systems that enable interactive/online code execution and commenting of the code blocks. This topic is still debated in the community, and there are dissenting opinions on "reproducible research" itself such as [https://doi.org/10.1080/0952813X.2017.1413140](https://doi.org/10.1080/0952813X.2017.1413140). Lepton supports / implements a proposal for reproducible research papers that is focused on plain text for long term readability and universality (not dedicated to a specific programming language). The two cited papers in ICCS are about Lepton. In addition, Lepton is designed to be easily re-implemented for long term maintainability ; publishing the documented implementation details provides a reference implementation and preserve the research.

Can I improve the paper.md file after submission following your comments ? How should I do it ?
