### 2018-07-28 #836 @jedbrown
1. How was `lepton.bin` bootstrapped? Is it possible to recreate without executing the existing Linux/x86-only binary?

2. Correct, CeCILL-B (which is very different from CeCILL-2.1) is not OSI-approved.

3. This is more "meta" than typical, but there have been accepted papers on tooling to facilitate reproducible research and I think it would be in scope according to that precedent.
