### 2018-08-03 6681512
License modification to OSI-approved CeCILL-2.1
