### 2018-08-03 11e094f
Improve JOSS submission : - clearer research usage - specificity as a tool for reproducible research
