### 2018-08-03 #836 @slithiaote
1. How was lepton.bin bootstrapped? Is it possible to recreate without executing the existing Linux/x86-only binary?
-- From the point of view of the computer, lepton.bin is compiled from main.ml (and several modules such as lepton.ml, lexer.ml, interpreters.ml, see chunk 26 on page 8 of lepton.pdf)
-- It is possible to recreate the source files from lepton.nw, with just copy-and-paste. The executable does that for you, and executes the instructions in chunk 16 for lexer.ml and chunk 26 for compiling, but everything is manageable by hand. It is a feature of Lepton that the source code is not too large so that it should be easy to recreate by hand, and sufficiently documented  so that it should be easy to re-implement for long-term maintainability.
-- Bootstrapping starts with a program that assembles and writes chunks to disk from a .nw file. As soon as you have that, you can stop writing source code directly, and write lepton.nw instead.
-- The second stage is command execution via the shell interpreter, which enables lepton to generate itself by typing lepton.bin lepton.nw
-- A third stage is to encapsulate the source code in an executable program. Lepton does this partially. As lepton can process input on stdin, it is usable as a command interpreter via the #! mechanism. On a suitably configured system, the command ./lepton.nw in a terminal will generate the lepton.bin executable and the manual.
1. Correct, CeCILL-B (which is very different from CeCILL-2.1) is not OSI-approved.
-- I understand. I will switch to CeCILL-2.1 then.

I have updated the github repository and improved the paper.md file following your comments. Was that the correct way to proceed ?
