### 2018-08-03 #836 @jedbrown
1. Thanks for the explanation. Please consider the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).

2. You could consider an OSI-approved permissive license. Changing to CeCILL-2.1 is far more consequential.

3. Yes, updating the repository to improve `paper.md` and/or address reviewer comments is the intended process.
