### 2018-08-30 53f5cfc
Better instructions for "Obtaining the software".
