+++ biblio_lepton.bib
@@ -1,6 +1,3 @@
-% This file was created with JabRef 2.7b.
-% Encoding: UTF8
-
 @MANUAL{Pygments,
   title = {Pygments},
   author = {Georg Brandl and Tim Hatch and Armin Ronacher},
@@ -80,4 +77,3 @@
   doi          = {10.5281/zenodo.1311588},
   url          = {https://doi.org/10.5281/zenodo.1311588}
 }
-
