### 2018-08-30 #836 @slithiaote
Following your comments, I have modified the "Compilation" section into broader and more detailed instructions for "Obtaining the software".
I have re-read and I hope that this should satisfy the review criteria. If not, could you be more specific as to what you feel needs improvements ?
