### 2018-09-08 #836 @kit-ty-kate
I'm not sure I'm well suited to review this.
I tried to compile the project, there were no instructions to compile anything and the only thing I could try (i.e. `./lepton.bin lepton.nw`) ended up in an infinite loop with some error messages and some OCaml files being generated.
The generated code itself is not very readable, almost no indentation and I don't know what to do with it. Also it uses old-style OCaml constructs but that's minor. I couldn't find any instructions to build the generated OCaml files but I managed to have something with `ocamlbuild -package str,unix main.byte` in the end (but it produced some maybe relevant and some irrelevant warnings)

Modern OCaml code have at least an opam file and it's not the case here, but I guess it's maybe not the goal of the tool as it seems to have its own build format (e.g. `<<shell_compile -exec shell>>=` in `lepton.nw`)
I'm certainly not an expert in literate programming and don't know much about it so I can't really comment on that I guess apart from the fact that it is very different from how OCaml projects are usually handled.
