### 2018-09-10 #836 @arfon
> The article proof does not render the item list, whereas the github preview does.

@slithiaote - this PR should fix the rendering issue in the PDF: https://github.com/slithiaote/lepton/pull/1
