### 2018-09-10 #836 @yochannah
@slithiaote thanks for the edits you've made. I think there are still some formatting issues in your pdf around the quotes - it looks like maybe there are two single apostrophes `''` rather than a single double quote `"` around Lepton, which makes the proof look a little strange.

I'm still hunting for reviewers & will update when I find someone.
