### 2018-09-13 #836 @adammichaelwood
I'm available for review. Let me know how I can best help.
