### 2018-09-13 #836 @yochannah
@adammichaelwood Hooray, and thank you for volunteering!

I would like to have a second reviewer for this paper. I'll kick off the official reviewer process once I find a second (if you can suggest anyone who might be able to review this, that would be awesome!). In the meantime feel free to look at the package and look at the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) for JOSS.
