### 2018-10-08 #836 @tpetricek
Hi all, @yochannah asked me to be the second reviewer for this - I'm happy to accept. I'll go ahead and put together a review following the [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) in the next few days. Is there anything else that I need to do?
