### 2018-10-23 #1005 @tpetricek
Thanks for the nudge! I was at a conference last week, but will post my initial comments today or on Thursday.
