### 2018-10-26 slithiaote/lepton#2 @tpetricek
Compiling lepton from scratch
My understanding is that `lepton.nw` is the implementation of Lepton in itself which, when processed through Lepton, will produce a `tex` file with description and an OCaml source code that can then be compiled to get Lepton running. Is that right?

However, I could not figure out how to actually bootstrap the process. The repository has `lepton.bin` binary, but I don't know how to run this (neither on Windows, nor using the Ubuntu bash on Windows) and I also think it's wrong to have binary in the repo.

I think it would be better if you included OCaml source code for bootstrapping, perhaps with `Makefile` or some other easy to run script that will produce a compile binary for my current platform. Then I could run that to process `lepton.mw` and get my own build. However, given the current setup, I was not able to run anything.
