### 2018-10-26 #1005 @tpetricek
_[Better late than never! Below is my review. I was not able to tick all the boxes above, so the review explains what problems I had and what I think needs to be corrected. I opened one issue for Lepton tracking the most important question, i.e. [how to actually compile it](https://github.com/slithiaote/lepton/issues/2)]_

The project is a literate programming package that allows users to write code in a classic literate programming style as proposed by Knuth. In case of Lepton, this means using LaTeX syntax with embedded chunks of code that can be combined together to define the program (this also means you can refer to a code block defined later in the file). Lepton is mostly an implementation of the original idea by Knuth, without adding much new (although the documentation is not very clear about that, so there might be new things I might have missed). However, that seems reasonable - a good implementation of the idea for OCaml would be a useful contribution.

Lepton is "implemented in itself" and my understanding is that to build it, you first need to run an existing version (to bootstrap it) and to obtain OCaml code which you can then compile to actually compile Lepton. However, I was not able to do this from scratch (see issue https://github.com/slithiaote/lepton/issues/2). I think the repository needs a Makefile of some sorts and clear instructions about how to bootstrap things (possibly for different OSes, if this cannot be universal).

Not being able to bootstrap Lepton meant that I was not able to tick the following review boxes:

 - Installation - as discussed above, I'm not sure how to bootstrap the build process
 - Functionality - consequently, I was not able to run Lepton and see what it does

Out of the documentation required for JOSS, I think the following are missing or not clear:

 - Installation instructions - I understand that this is tricky given the nature of the project (and the fact that it is written using itself), but there should be clear  instructions how to compile it (for some supported platform) and run it. I think this should also use OCaml source code rather than binary for bootstrapping.

 - Example usage - There are some instructions in `lepton_manual.pdf`, but it is not entirely clear to me how to go from `nw` file to an executable (how to go from `nw` file to LaTeX file is however quite clear).

 - Automated tests - As far as I can see, there are no tests

 - Community guidelines - Besides the standard license file, which sets the basic rules, there is not much to guide community contributions. Having a standard CONTRIBUTING file would be useful.
