### 2018-11-04 #1005 @yochannah
@slithiaote let me know if/when you've addressed the relevant issues and are ready to move forward with this
