### 2018-11-07 slithiaote/lepton#2 @slithiaote
Thank you for your comments

Your understanding is mostly correct. When processed through Lepton, lepton.nw will produce the OCaml source code **and** compile it. You should obtain a new binary. As such, lepton.nw is in itself a kind of Makefile or script that is easy to run, albeit you need a first implementation of Lepton.

The executable is for linux systems and should work on 32bit and 64bit systems. It won't work on Windows. I am unsure about Ubuntu bash, but I believe it should work if you write the command in a terminal, or at least give you a useful error. Were you able to run the command `./lepton.bin lepton.nw -o lepton.tex` in Section 3 of `lepton.pdf` (the rendered version of `lepton.nw`) ?

While I agree that distributing binaries is wrong in general, I think that Lepton defines a "new source code format", and is similar to compilers such as gcc in that respect. As far as I understand, you have to provide a binary implementation of the compiler in order to compile its source code.
