### 2018-12-03 #1005 @slithiaote
@yochannah I replied to the issue/report on Nov 7th. At least part of the problem seems to be Windows, which is not a supported platform. I have asked for more input, but @tpetricek did not reply and I am hesitant to ping.
