### 2018-12-05 slithiaote/lepton#2 @tpetricek
I get the following when I try to run the above command:

```
root@lipwig:/mnt/c/Tomas/Public/tpetricek/lepton# chmod +x lepton.bin
root@lipwig:/mnt/c/Tomas/Public/tpetricek/lepton# ./lepton.bin
-bash: ./lepton.bin: cannot execute binary file: Exec format error
```

That said, I think you should be able to avoid distributing binaries. If you provide the OCaml source code generated from your `nw` source file, together with a Makefile, then I think that would make the bootstrapping a lot easier (and it would enable you to support multiple platforms).
