### 2018-12-05 #1005 @tpetricek
@slithiaote I missed the reply. Sorry. I just added a comment on the issue.

That said, I think there are other issues (see my comment above) that need to be addressed before this can be accepted as a JOSS publication - there are a few things that the JOSS checklist requires that I think are not yet satisfied by lepton.
