### 2018-12-14 #1005 @yochannah
@tpetricek Sorry, I missed the edit note. I can look for another reviewer - it took me a while to find people interested in literate programming - you don't have any suggestions of possible interested colleagues or communities to approach?
