### 2018-12-16 #1005 @tpetricek
@yochannah I'm happy to stay involved if the author is interested in supporting Windows, at least through the Ubuntu Bash shell - I don't see why that should be impossible and I think it would be a good thing for the project in general.

No matter what the supported platforms are, I think the project should use some form of pre-generated OCaml source code for bootstrapping rather than a checked-in executable where nobody knows what the executable might be doing. I think making that change (and adding a makefile) would likely make this work on Windows anyway.

That said, if the project can be accepted with a binary file in the repository, then it needs to be reviewed by someone who is able (and brave enough) to run it, so in that case, I won't be able to help. I sadly don't have any good alternative reviewer ideas.
