### 2018-12-17 #1005 @slithiaote
@tpetricek Lepton compiles on Windows, but is mostly unusable because it requires the Unix library and Unix tools. Your comment made me realize that Windows Subsystem for Linux can provide those, although I will not be able to test it myself. I have been working on a source-based distribution for the past few days. I should be able to update this week.

@yochannah Have you tried the "reproducible research" community ?
