### 2018-12-19 #1005 @yochannah
Okay - this is good - it sounds like we may be able to keep @tpetricek onboard if we can get this on Windows! 👍

I do think I generally agree about @tpetricek's comment re pre-generated OCaml source code.
