+++ biblio_lepton.bib
@@ -1,3 +1,4 @@
+@Comment This file was generated from lepton.nw by lepton.bin
 @MANUAL{Pygments,
   title = {Pygments},
   author = {Georg Brandl and Tim Hatch and Armin Ronacher},
