+++ Licence_CeCILL_V2.1-en.txt
@@ -20,7 +20,7 @@ the two main principles guiding its drafting:
 The authors of the CeCILL (for Ce[a] C[nrs] I[nria] L[ogiciel] L[ibre]) 
 license are: 
 
-Commissariat à l'énergie atomique et aux énergies alternatives - CEA, a
+Commissariat Ã  l'Ã‰nergie atomique et aux Ã©nergies alternatives - CEA, a
 public scientific, technical and industrial research establishment,
 having its principal place of business at 25 rue Leblanc, immeuble Le
 Ponant D, 75015 Paris, France.
@@ -341,7 +341,7 @@ The Licensee expressly undertakes:
 
 The Licensee undertakes not to directly or indirectly infringe the
 intellectual property rights on the Software of the Holder and/or
-Contributors, and to take, where applicable, vis-à-vis its staff, any
+Contributors, and to take, where applicable, vis-Ã -vis its staff, any
 and all measures required to ensure respect of said intellectual
 property rights of the Holder and/or Contributors.
 
