### 2018-12-20 slithiaote/lepton#2 @slithiaote
I believe that today's commit should fix this issue, I switched to a source-based distribution.

I initially thought that command interpretation with Lepton could only be available on Linux with the current code base, so distributing a linux binary was acceptable. The possibility to use Lepton on Windows Subsystem for Linux tips the balance in favor of compiling from source.
