### 2018-12-20 #1005 @slithiaote
I just committed to Github and tagged v1.1
