### 2019-01-13 #1005 @yochannah
Hey all - just checking the status of this review. @tpetricek, has @slithiaote's update managed to resolve your issues? @adammichaelwood, do you think you might have a chance to complete your review soon? Thanks all!!
