### 2019-01-16 slithiaote/lepton#2 @tpetricek
Looking at the current version, I'm much happier about the new approach!

That said, just running `make.sh` does not yet work for me - presumably, because I need to install some dependencies. Is this documented somewhere?
