### 2019-01-16 #1005 @tpetricek
Thanks for the ping. I'm still trying to get it to build. Just added a comment there: https://github.com/slithiaote/lepton/issues/2

In addition to that, the project still does not satisfy some of the JOSS items from the check-list above, so adding those (assuming they are strictly required for acceptance) would be another good thing to do in the meantime (especially, tests, contributor guidelines)
