### 2019-01-25 slithiaote/lepton#2 @slithiaote
Apart from a standard installation of the Ocaml compiler, there are no dependencies for running make.sh. No external libraries are required.
Can you check that `ocamllex` and `ocamlopt` are installed ? I just recall that some Linux distributions do not install the full OCaml system by default.
