### 2019-01-25 slithiaote/lepton#2 @tpetricek
No, I don't think I have OCaml installed - but I think that instructions like these should be included in the project readme or building guide.
