### 2019-01-28 slithiaote/lepton#2 @slithiaote
Section "3 Obtaining the software", page 8 of the `lepton.pdf` documentation describes how to compile the source files with Ocaml to obtain the executable, as well as the expected output of `make.sh`. The bootstrapping instructions are provided in the same section.
Are these insufficient in your opinion ?
