### 2019-01-28 slithiaote/lepton#2 @tpetricek
I think these are sufficient for a personal project, but not for a high-quality easy to use open source software published in JOSS. I just had a quick look at the last 5 projects accepted to JOSS and they all either have instructions directly in README or they have a direct link to documentation which has installation instructions.
