### 2019-02-13 bb0a8b2
add links to the pygments and LaTeX install pages
this way people setting up Lepton can more easily get the dependencies
