### 2019-02-13 slithiaote/lepton#2 @yochannah
@slithiaote I note you've made a nice readme update. I have a few comments:

- I installed ocaml and tried run the [setup script](https://github.com/slithiaote/lepton#installation). Is this right?

```bash
chmod +x make.sh
./make.sh
```

This output the following to the console. I'm not sure if it worked?

```bash
101 states, 393 transitions, table size 2178 bytes
1974 additional bytes used for bindings
File "lexer.mll", line 73, characters 20-35:
Warning 10: this expression should have type unit.
File "lexer.mll", line 42, characters 57-58:
Warning 26: unused variable s.
File "lexer.mll", line 43, characters 63-64:
Warning 26: unused variable s.
File "lexer.mll", line 42, characters 33-34:
Warning 26: unused variable h.
File "lexer.mll", line 42, characters 33-34:
Warning 26: unused variable h.
File "lexer.mll", line 42, characters 33-34:
Warning 26: unused variable h.
File "lexer.mll", line 42, characters 57-58:
Warning 26: unused variable s.
File "lexer.mll", line 45, characters 27-28:
Warning 26: unused variable b.
File "lexer.mll", line 45, characters 68-69:
Warning 26: unused variable s.
```

When I run `lepton` in the console after this, my console says `command not found: lepton`

Any suggestions - am I doing something wrong?

I'm still working on trying the bootstrapping instructions and may have further comments :)
