### 2019-02-13 #1005 @yochannah
Quick note the readme on this repo was updated recently! :) I've added some comments about build issues in https://github.com/slithiaote/lepton/issues/2
