### 2019-02-14 #1005 @slithiaote
Yes, I updated the readme recently, but I did not have the time to comment as well. As far as I understand, @tpetricek expects to find installation instructions in `README.md` instead of `lepton.pdf`, so I included a copy of these instructions in `README.md`, as well as contributor guidelines similar to what is present in a few recent accepted JOSS papers.

The project contains two test files: an embedded `lepton_manual.nw` to produce the usage manual and `lepton.nw` itself. These two files use most of the functionality of Lepton. I believe that bootstrapping Lepton constitutes a fairly complete test of the software. Incidentally, boostrapping will automatically attempt to process `lepton_manual.nw`.
