### 2019-02-17 #1005 @yochannah
@slithiaote I think I generally do agree - clear getting started docs in the readme is important for all open source software. I've added some comments myself in https://github.com/slithiaote/lepton/issues/2 as I also had trouble getting things to run.
