### 2019-02-19 b301052
Update to the documentation - show expected warnings in lepton.pdf - renaming the executable is necessary for bootstrapping
