### 2019-02-19 slithiaote/lepton#2 @slithiaote
The warnings are expected, I believe the pattern-matching code is easier to read with them.

`make.sh` generates `lepton.bin`, so you should run `./lepton.bin` if you want to use the program. For bootstrapping, you should rename the executable. I have just updated the documentation and tested on my computer.
