### 2019-03-29 #1005 @tpetricek
@slithiaote Going back to the checklist above, there are two more items "Automated tests" and "Community guidelines" that I did not see in the repo - assuming those are required by the JOSS, would you be able to add those?
