### 2019-04-04 #1005 @slithiaote
@tpetricek I believe that "Automated tests" is covered by "Bootstrapping", as this exploits the functionality for which Lepton was written. Bootstrapping will in fact process two embedded files, i.e. run two test files.

I included "Community guidelines" in the "Contributing" section in `Readme.md`, as this is what I found in other recently published JOSS papers.
