### 2019-05-11 #1005 @labarba
JOSS guidelines say this about [tests](https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=tests#tests):

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
> - Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> - OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> - Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works
