### 2019-05-15 #1005 @slithiaote
I believe that the bootstrapping process falls into the OK category. The process is manual, and the steps are documented. This process will run two test files which cover the expected functionality of Lepton :
- it will write files to disk, e.g. the source code
- it will run commands, e.g. the compilation commands to produce the Lepton executable
- it will produce documents, e.g. the sources for the manual and tutorial, which can then be compiled to PDF with LaTeX (instructions included)
The resulting files can be objectively checked against the ones provided in the distribution. In particular, the generated source files should be identical.
