### 2019-06-10 #1005 @danielskatz
👋 @tpetricek - please confirm that the documentation for testing is OK, given the discussion above, and if so, please check off the item. If you think something else is needed to make this OK, please explain what you believe is needed.

And for other items in the checklist, what are the blockers to you?
