### 2019-06-10 #1005 @tpetricek
Hi @danielskatz - unfortunately, I no longer have access to a UNIX machine in order to be able to run the software, so I'm unfortunately not able to check the functionality claims.
