### 2019-06-18 #1005 @kyleniemeyer
Hi @slithiaote @adammichaelwood and @tpetricek , since @yochannah is stepping down from the JOSS editorial team, I'm going to take over editing this article moving forward. Thanks @yochannah for shepherding it so far!
