### 2019-06-18 #1005 @kyleniemeyer
It sounds like @adammichaelwood might be able to complete his review in the near future—thank you! I'll set a reminder.

@tpetricek I will check the functionality claims myself, if you are willing to complete the rest of your review.
