### 2019-06-18 #1005 @tpetricek
I think the remaining items really depend on the JOSS standards and I don't have a good sense about those. In particular:

- Example usage - The repo does have one example, which is self-hosting. Technically, that's an example usage, but it's pretty sophisticated one that does not actually teach us much about how anyone might use the system for real work. It checks the box "in principle", but I'm not sure it's a good example.

- Automated tests - Technically, self-hosting is a test - but again, this is not what I'd expect. I'd expect some automated unit testing, perhaps running using continuous integration. This is technically ticking the box, but again, I don't think it's following a very high quality standard.

- Community guidelines - There is one vague sentence in the README. This is technically a "community guideline" but it is very basic one, e.g. there is no `CONTRIBUTING.md` file and it does not really guide anyone interested in how they might get started contributing (e.g. no `up-for-grabs` issues).

So, I feel that someone with better idea about JOSS standards should decide. I personally don't think the project is "quite there" but maybe my standards are more strict than they should be...
