### 2019-06-24 #1005 @kyleniemeyer
Hi @slithiaote, all of @tpetricek's comments above are valid and appropriate from my point of view. Can you address or respond to each?
