### 2019-07-01 #1005 @kyleniemeyer
@slithiaote regarding your responses,

1. There may technically be two examples, but at least one reviewer found this a bit lacking. Could you improve the documentation around the examples, to make it easier for a new user? It may make sense and work easily for you, but both @yochannah and @tpetricek ran into issues.

2. All software changes, and research software needs tests. JOSS publishes research software, and some sort of automated test is a requirement.

3. I agree that the "contributing guidelines" is extremely minimal. Also, in fact problems have been reported, in an issue filed by @yochannah and @tpetricek.

I will add that in general, we expect that authors of submissions to JOSS will use the peer-review feedback to improve the software—this is one major point of the journal. So far, you have generally been rebutting the feedback, rather than improving your software / documentation / etc. In fact, @yochannah and @tpetricek even filed an issue on the repo, which you haven't satisfactorily addressed, and your repo has not been updated in months.

Most JOSS submissions are not published in the state that they were submitted. A few issues have been raised here, and we'd expect that you would take that useful feedback in the spirit it was given, and work to improve your submission.
