### 2019-07-01 #1005 @slithiaote
1. Both @yochannah and @tpetricek ran into installation issues, which I have addressed by switching to a source-based distribution to support Windows Subsystem for Linux, adding a `make.sh` script (3 lines of code), and writing more detailed instructions and putting those in the `README`, instead of the main documentation file `lepton.pdf`. I have replied to the comments on the github issue, but I feel that it is good practice to wait for an acknowledgement that the issue is resolved for the user before closing the issue, so it is still marked as open.
I am happy to add an intermediate example, but I will need a few days to design one that fits the "real work" criterion. I still believe that `lepton.nw` is not too sophisticated for that.

2. I agree that all software need tests. In the case of Lepton, I believe that bootstrapping is the best test, and the comments have not indicated flaws with this. Let me point out that boostrapping runs Lepton on a sophisticated example, and writes its output to disk. The command `git commit` will automatically run the `diff` command on the generated source files and check that these correspond exactly or show the modifications. It is the purpose of Lepton to generate these `.ml` and `.tex` files. I don't understand why this does not qualify as testing, or what benefits a continuous integration system would provide.

3. The "contributing guidelines" are taken from published JOSS papers. They are minimal, but Lepton is not such a large scale project as to require procedures beyond "write an e-mail" or "open an issue on Github". I'd rather keep this simple and welcoming.

I do appreciate feedback, and I have taken care and time in answering. On the topic of installation (open issue in the repository), I have addressed the issue and am still waiting for a followup. On the topic of "automated tests", I have replied to the comments (I claim the right to disagree), but again without a discussion. This has not been helpful to assess whether the issue is satisfactorily addressed or not and why.
