### 2019-07-11 ee3c3bb
Initial commit of fibonacci.nw example. Update the Readme file to document this.
