+++ README.md
@@ -46,6 +46,12 @@ The `lepton.bin` executable can be compiled with the included `make.sh` script.
 Requirements
 - [OCaml](https://ocaml.org/docs/install.html) > 4.0
 
+# Examples
+Three examples are provided :
+- `hello.nw` is a minimal working example showing how to embed OCaml source code and execute it. It requires only LaTeX and OCaml. See Section 1 of `lepton_manual.pdf` for details.
+- `fibonacci.nw` is an example of a scientific report comparing several implementations of the Fibonacci sequence and their running times. The corresponding rendered PDF `fibonacci.pdf` contains instructions on how to execute it. It is currently a "work-in-progress".
+- `lepton.nw` is the source file used for the source code and the documentation of Lepton. Executing this example will regenerate both the binary executable and the PDF documentation. The process is described in the "Bootstrapping" section of this document and in `lepton.pdf`.
+
 # Bootstrapping
 The contents of the installation package can be re-generated from the file `lepton.nw`. This will produce a new executable, extract copies of the source files, and compile the LaTeX documentation.
 
