### 2019-07-12 #1005 @kyleniemeyer
@slithiaote I appreciate your response, and apologies for the delay here—the original second reviewer was unable to complete their review, and so I had to find someone else.

Fortunately, @rljacobson has agreed to review. Thank you!

Apologies again for how long this is all taking; I'm hoping @rljacobson's review sheds light on any final changes needed, particularly for ensuring new/unfamiliar users can build and use the software.
