### 2019-07-13 #1005 @slithiaote
I uploaded an initial draft of an additional example. It shows how to use Lepton to write a reproducible scientific report. Here, reproducible means that all the source code used in the report is provided, as well as data files, instructions for producing the figures. The `.tex` report itself can be re-generated with a single shell command, and then compiled to PDF with LaTeX.
