### 2019-07-22 87276a6
Improve fibonacci.nw example
