### 2019-07-22 #1005 @slithiaote
I uploaded an improved version of the Fibonacci example. It is now feature-complete and is ready for review.
