### 2019-07-30 #1005 @holtgrewe
Hi, after reaching out to @kyleniemeyer, here are some points **outside** a formal review (if only out of sentiment from my original background, I'd like to see some more support for the literate programming style in any language):

- Given that Markdown/HTML are stronger on Github that classic LaTeX, it's a bit of a pity that all of the relevant information hides in binary PDF rather than text formats.
- Also, the front-page `README.md` file did not speak out to me as "Hi, I'm the Markdown that you're used to on a repository front page but I'm a relatively involved MD document with some YAML on the top that you're not interested in." Maybe this could be replace with a common MD document showing links to what people like: continous integration, e.g., through Travis or CircleCI?
- Maybe the PDFs could be built in continuous integration (maybe even HTML?) and be pushed into a separate branch to highlight with "living examples" that this tool can be used on modern infrastructures and setting it up is reproducible?

Although being a big fan of cameloids 🐫 🐪 and having some exposure to Haskell and the like, I'm not an OCaml expert. I'd be happy to try out all the examples and generatin all the PDFs. I'd be happy to install all the packages on all the recommended OSes, but I'm missing the necessary steps.

Maybe the author could provide the steps to install the dependencies on one popular Unixoid operating system (Ubuntu/Debian/CentOS, I'd be happy to start off FreeBSD). If you could provide step-by-step instructions in recreating the PDFs in the repository (starting off `apt-get/yum install` of your wonderful OCaml packages, continuing with `git clone` and then calling any build scripts/makefile etc.), I'd be happy to follow along.

Ayway, this is just as an "voice from the off", feel free to ignore me.
