### 2019-08-06 #1005 @slithiaote
I agree that it would be more natural to write in Markdown for Github, and this is possible in Lepton. However, my priority when writing Lepton was scientific journals, hence LaTeX. Incidentally, the relevant information hides in the `lepton.nw` file, which is a LaTeX / text format, and not so much harder to read than Markdown, at least for this document.

Concerning README.md, I tried to mimick JOSS examples in the guidelines, but I can very well be wrong there.

Lepton's mindset is to let users reproduce things themselves or show them everything they need. I'd rather users do the steps instead of having a simple command to clone a prebuilt setup. I hope this will make it possible to build Lepton ten years from now, with small adjustments, and I believe that it is not a big hurdle here.

Here are some instructions for Debian
- OCaml can be installed with `apt-get install ocaml`.
- The git clone command is `git clone https://github.com/slithiaote/lepton.git`
- You can then compile Lepton with the `make.sh` script in the repository (run `sh make.sh`).
- Regenerating the PDFs will require LaTeX : `sudo apt-get install texlive python-pygments` should install everything. In particular, the `minted` package is included in `texlive-latex-extra`
