### 2019-09-11 #1005 @Kevin-Mattheus-Moerman
@tpetricek thanks for helping us with this lengthy review process! Can you please list/summarize the main points you feel @slithiaote should focus on? Also can you please comment on (apologies if I've missed your comments somewhere) this reply by @slithiaote:

> I think that @tpetricek's comments are more appropriate for a large scale software project with ongoing development, whereas this submission is intended as a reference implementation for long-term preservation of the software.
>
>     * Example usage - Technically, there are two examples. A very simple `tutorial.nw` and the full-fledged `lepton.nw`. As in the literate programming paradigm, Lepton is intended to produce documents containing source code and executable instructions. It was used to produce this submission, so `lepton.nw` is a good example of real work in my opinion
>
>     * Automated tests - I think that continuous integration is suited for projects that evolve rapidly. As it is intended as a reference implementation, the source code of Lepton has not evolved in months. Moreover, testing can be done with `diff` to make sure that the generated `.ml` and `.tex` files are the same. In fact, this is performed automatically by `git commit`.
>
>     * Community guidelines - no problems have been reported for months, so there are no "up-for-grabs" issues at the moment.
