### 2019-10-15 #1005 @rljacobson
@kyleniemeyer  My review is complete, but it appears I cannot edit the checklist. I have one issue:

The paper (paper.md) has an empty References section. That section should presumably include the references `@LiThiaoTe20121723` and `@LiThiaoTe2012439`, and possibly `@Knuth84literateprogramming` as well.

# Review checklist for @rljacobson

## Conflict of interest

- [X]   As the reviewer I confirm that I have read the JOSS conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [X]  I confirm that I read and will adhere to the JOSS code of conduct.

- [X] Repository: Is the source code for this software available at the repository url?
- [X] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [X] Version: Does the release version given match the GitHub release (v1.0)?

- [X]     Authorship: Has the submitting author (@slithiaote) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [X] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

- [X] Authors: Does the paper.md file include a list of authors with their affiliations?
- [ ] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
