### 2019-10-15 #1005 @kyleniemeyer
@rljacobson thanks! The references are automatically included in the PDF from the associated `.bib` file, and aren't supposed to be in the Markdown file, so this looks ok.
