### 2019-10-15 #1005 @rljacobson
@kyleniemeyer Great, then my review is complete. He gets a pass from me. Looks good.
