### 2019-10-15 7c2f10a
Update paper.md
Some minor fixes to the JOSS paper
