### 2019-10-15 #1005 @kyleniemeyer
@slithiaote at this point I believe we can accept this submission. I just submitted a minor PR with fixes to the paper, if you could merge this: https://github.com/slithiaote/lepton/pull/4

Then, please archive the software repository at Zenodo and let me know the resulting DOI.
