### 2019-10-16 10f673a
Add DOIs to references and some minor LaTeX formatting
