+++ biblio_lepton.bib
@@ -8,13 +8,21 @@
 }
 
 @ARTICLE{Knuth84literateprogramming,
-  author = {Donald E. Knuth},
-  title = {Literate programming},
-  journal = {THE COMPUTER JOURNAL},
-  year = {1984},
-  volume = {27},
-  pages = {97--111}
+    author = {Knuth, D. E.},
+    title = "{Literate Programming}",
+    journal = {The Computer Journal},
+    volume = {27},
+    number = {2},
+    pages = {97-111},
+    year = {1984},
+    month = {01},
+    abstract = "{The author and his associates have been experimenting for the past several years with a programming language and documentation system called WEB. This paper presents WEB by example, and discusses why the new system appears to be an improvement over previous ones.}",
+    issn = {0010-4620},
+    doi = {10.1093/comjnl/27.2.97},
+    url = {https://doi.org/10.1093/comjnl/27.2.97},
+    eprint = {http://oup.prod.sis.lan/comjnl/article-pdf/27/2/97/981657/270097.pdf},
 }
+
 
 @MANUAL{ocamlrefman,
   title = {The OCaml system (release 3.12): Documentation and user's manual},
@@ -60,15 +68,24 @@
 }
 
 @ARTICLE{ramsey1994literate,
-  author = {Ramsey, N.},
-  title = {Literate programming simplified},
-  journal = {Software, IEEE},
-  year = {1994},
-  volume = {11},
-  pages = {97--105},
-  number = {5},
-  publisher = {IEEE}
-}
+ author = {Ramsey, Norman},
+ title = {Literate Programming Simplified},
+ journal = {IEEE Softw.},
+ issue_date = {September 1994},
+ volume = {11},
+ number = {5},
+ month = sep,
+ year = {1994},
+ issn = {0740-7459},
+ pages = {97--105},
+ numpages = {9},
+ url = {https://doi.org/10.1109/52.311070},
+ doi = {10.1109/52.311070},
+ acmid = {625384},
+ publisher = {IEEE Computer Society Press},
+ address = {Los Alamitos, CA, USA},
+ keywords = {code extraction, documentation, feature complexity, language-dependence, literate programming, literate programming tools, programming, programming environments, software tools, source file, system documentation, tool acceptance},
+} 
 
 @misc{zenodo,
   author       = {Li-Thiao-T\'{e}, S\'{e}bastien},
