### 2019-10-16 #1005 @slithiaote
I just tagged the release v1.2 with DOI 10.5281/zenodo.3492221
I added missing DOIs to the bib file and made cosmetic LaTeX changes.
Thank you for the suggestions.
