### 2019-10-17 #1005 @kyleniemeyer
@slithiaote sorry, one last thing: can you clean up the metadata on the Zenodo archive? The title and author list should match the paper.
