### 2019-10-18 #1005 @slithiaote
Is it perhaps related to the comment in the first line of my bibtex file ?
`@Comment This file was generated from lepton.nw by lepton.bin`
