### 2019-10-21 eed4ddf
Removed comment line to aid JOSS paper
