### 2019-10-22 #1005 @slithiaote
I would like to keep the Bibtex comment because it makes it clear that users should not edit the bib file directly. I can remove `@COMMENT` and leave the text dangling outside the Bibtex entries. Both methods are official ways to add comments in Bibtex files.

Do I need to tag a new release on zenodo with tag `@openjournals/joss-eics` ?
