### 2019-10-22 #1005 @arfon
> I can remove @COMMENT and leave the text dangling outside the Bibtex entries. Both methods are official ways to add comments in Bibtex files.

That probably works. No new release is necessary.
