### 2019-10-22 13ae9fd
Remove at-sign for Bibtex comments
