### 2019-10-24 #1005 @Kevin-Mattheus-Moerman
@openjournals/dev this DOI is not resolving yet
