### 2019-10-24 #1005 @arfon
> @openjournals/dev this DOI is not resolving yet

Looks like the Crossref queue is backed up. It should register in the next few hours.
