### 2019-10-24 #1005 @kyleniemeyer
Congrats @slithiaote on your article's publication in JOSS—thanks for being patient over this long process. Many thanks to @tpetricek and @rljacobson for reviewing this submission!
