### 2019-10-28 #1005 @slithiaote
Thanks to the editors and reviewers from JOSS for their dedication and perseverance.
